[BUGFIX] Page cache expiry calculation fails
The calculation of the page cache expirationy timestamp based on
property config.cache.xx = table:yy fails in some specific
scenarios, name when there are records having either a start time
in the future but no end time, or an end time in the future
and no start time. Contrary to what is being done so far, calculation
must be performed separately for each time field to avoid
interferences.

Change-Id: I39f08a84255233952bd2fd51f28972bb1fe9affa
Fixes: #35684
Releases: 6.0, 4.7, 4.6
Reviewed-on: http://review.typo3.org/11095
Reviewed-by: Francois Suter
Tested-by: Francois Suter