[BUGFIX] Page cache expiry calculation fails
The calculation of the page cache expirationy timestamp based on
property config.cache.xx = table:yy fails in some specific
scenarios, name when there are records having either a start time
in the future but no end time, or an end time in the future
and no start time. Contrary to what is being done so far, calculation
must be performed separately for each time field to avoid
interferences.

Change-Id: I580cb6ca02c6efd9b594b31bd887cff9b9b3147c
Fixes: #35684
Releases: 6.0, 4.7, 4.6