[FEATURE] rsaauth calls openssl_pkey_new on every request
openssl_pkey_new is not a simple cheap operation and for minimal (ajax) requests this can consume 10% of the request time.

Currently its called in every request as part of the tx_rsaauth_php_backend->isAvailable() method.
I think its better to remove the call here and assume it works (which is the likely case) if this method is available.
The unlikely error is thrown when its processed - and the user knows that something is wrong with PHP openssl config.

Change-Id: I34eae9978a8cb0ff2b985a23a358dc88f55ab171
Resolves: #35798
Releases: 6.0