[BUGFIX] rsaauth PHP backend calls key generation on every request
Problem: openssl_pkey_new is not a simple cheap operation and 
for minimal requests (e.g. ajax), this can consume up to 10% of the request time.

Solution: Do not process the login fields if no login is triggered.

This commit will save even more processing power as the whole service chain
for processing the login fields is not instanciated any more.


Change-Id: I34eae9978a8cb0ff2b985a23a358dc88f55ab171
Fixes: #33165
Releases: 6.0, 4.7, 4.6