[BUGFIX] t3lib_div::getUrl() providing wrong error information
When using t3lib_div::getUrl() with lib file/context it uses
file_get_contents() to read a URL. When this fails it tries to provide the
reason for the error but does not return the actual issue.

The error key in the returned array contains the error retrieved by
error_get_last() which does not provide the error that occurred with
file_get_contents(), but with PHP. Usually this is a notice.

The actual error message of file_get_contents() is in the variable
$http_response_header which is available after calling file_get_contents().

See http://de2.php.net/manual/en/reserved.variables.httpresponseheader.php

Change-Id: Iddb421b53d20b03fed2fb765e5643d1e95e7da98
Fixes: #35847
Releases: 6.0, 4.7, 4.6, 4.5