[TASK] Streamline disposal of Signal Slot
The concept of Signal Slot has been introduced into Extbase of
TYPO3 4.6 and then got copied to TYPO3 4.7 as well - basically
to support that concept for the File Abstraction Layer as well.
However, now this observer pattern is duplicated in the Core,
thus is removed in favour of the available Extbase behaviour.

Change-Id: I406a2381771e91c119240347d0d0444844a2fc0a
Releases: 6.0, 4.7
Resolves: #35854
Reviewed-on: http://review.typo3.org/10252
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader