[TASK] Streamline disposal of Signal Slot
The concept of Signal Slot has been introduced into Extbase of
TYPO3 4.6 and then got copied to TYPO3 4.7 as well - basically
to support that concept for the File Abstraction Layer as well.
However, now this observer pattern is duplicated in the Core,
thus is removed in favour of the available Extbase behaviour.

Change-Id: I1b66cab4dbb5ee18a9e72df346a4b64711dd268a
Resolves: #35854