[BUGFIX] Centering images with text is not working in some cases
Currently css_styled_content is using the constant MaxW in CSS to define the width of the image wrap. When using columns with different widths, this value is not reflecting the width of the column, which can break the centering.

This patch fixes this issue by using a different centering method.

Change-Id: I2be4a53f010df8be7f7b0d5defd4bbbdad6a6e77
Fixes: #35814
Releases: 6.0, 4.7