[BUGFIX] Media Element does not play videos with own player anymore
Since the rewrite to FlowPlayer and the additions for accessibility
the MEDIA Content Element is not capable anymore, to play videos which
ship their own player.

This especially is the case for Videos which are from YouTube, Vimeo
and other similiar platforms. Looking at the implementation it seems,
that the detection of own players has been removed (accidentilly).

Rebuild this detection to make external video platform links work again.

Change-Id: Ifdc579b12794e09c51bd3ddad0290c2cc9470760
Fixes: #33357
Releases: 6.0, 4.7