[BUGFIX] Media Element does not play videos with own player anymore
Since the rewrite to FlowPlayer and the additions for accessibility
the MEDIA Content Element is not capable anymore, to play videos which
ship their own player.

This especially is the case for Videos which are from YouTube, Vimeo
and other similiar platforms. Looking at the implementation it seems,
that the detection of own players has been removed (accidentilly).

Rebuild this detection to make external video platform links work again.

Change-Id: I963a74aa3e1ab6d2de95ade903470be71af5c2d4
Fixes: #33357
Releases: 6.0, 4.7