[TASK] Remove doNotLoadInFE
The doNotLoadInFE handling was created to speed up the frontend by
not loading ext_tables files of extensions that are not relevant
in frontend context. In practice the effect is marginal, uglifies
the code adds not necessary context complexity. Furthermore the
TER was never able to handle the doNotLoadInFE value and removed it.
The whole handling is removed with the patch.

Change-Id: Icd68353bbb63decd08a86b686da3642f77317f40
Resolves: #35877
Releases: 6.0