[BUGFIX] Wrong HTTP headers sent when trying to access pages that require login
When trying to access a login protected page while not being logged in it
may happen that a wrong header is sent.
The install tool allows to set a header for the page not found
handling: [FE][pageNotFound_handling_statheader]
The default value for this setting is "HTTP/1.0 404 Not Found"
In a case where access is denied because the user is not logged in,
this is the wrong header, it should be 401 instead.

Change-Id: I6558fedc21ef3dfc203bccf2d39f8522f90109c6
Fixes: #23178
Releases: 6.0, 4.7