[FEATURE] New token for custom post configuraton in localconf.php
Some setups require to add a require() statement at the end of the
localconf.php to be able to override the configurations. This patch
introduces a new "edit end token" similar to the existing
"edit token". Only the lines between the "edit token" and the
"edit end token" will be updated by the install tool. The code above
and below that will stay untouched.

Change-Id: I87877503fe71196a7d099f96fb9f4a3018624bd5
Resolves: #24037
Releases: 6.0
Reviewed-on: http://review.typo3.org/10492
Reviewed-by: Tolleiv Nietsch
Tested-by: Tolleiv Nietsch