[BUGFIX] T3D export of tt_content elements with IRRE relations fails
Actually the import fails due to the fact that there is no informaton about the import action while calling process_datamap. So each of the process_datamap hooks will be called, even though they might be supposed to work with real save actions only.

Solution is to provide a boolean to indicate that TCE main has been called during an import action.

Change-Id: I5630b395d0c9bd2a01a5b4fa1a166ee3f863f400
Fixes: #36031
Releases: 4.5, 4.6, 4.7, 6.0