[BUGFIX] T3D export of tt_content elements with IRRE relations fails
Actually its the import that fails due to the fact that there is no information about the import action while calling process_datamap. So each of the process_datamap hooks will be called, even though they might be supposed to work with real save actions only.

Solution is to provide a boolean to indicate that TCE main has been called druing an import action.

Change-Id: I4fe429fb88f8ec24a7fda885f38ddfbdb4a642f5
Fixes: #36031
Releases: 4.5, 4.6, 4.7, 6.0