[BUGFIX] Remove XCLASS inclusion blocks from t3lib
After #31893 is merged we can remove the XCLASS blocks from the
Core. This patchset removes it from the classes within t3lib.

Besides that it declares t3lib_TimeTrackNull and t3lib_TimeTrack
to be final to make sure that extending them is still denied.

This patchset itself should not be merged before the related patch
for EXT:dbal (#31183) is merged.

Change-Id: I09e5253191972ac4b3b033aeb770758690ea3755
Fixes: #36184
Releases: 6.0