[!!!][FEATURE] Filename filters: pass additional information
The file/folder name filter methods now get an additional
parameter, $additionalInformation, that can be filled by
the driver as necessary. It could e.g. contain additional
information fetched from the storage, that might be relevant
for a filter.

This change requires adding this parameter in all filters, so it is
considered a breaking API change. Apart from that, driver authors have
to adjust their driver if they use the dispatch method from
AbstractDriver, as the parameter order changed there.

Change-Id: I24b43cc02c0bf79fb637b51081f65ef3321c5b99
Resolves: #36783
Releases: 6.0