[!!!][BUGFIX] RTE should honour FAL identifier in TCA special config
Problem: TCA special configuration options allow to specify an imgpath
as part of rte_transform option. If a FAL combined identifier is
specified as imgpath, the RTE will fail.

Solution:
Change the delimiter of RTEtsConfigParams query string from ":" to ';'.
This may break extensions that hook on browse_links and will most
probably break hooks on htmlArea RTE image and link dialogues by
extensions such as DAM.

t3lib_BEfunc::getSpecConfParts also needs to be adapted as it splits
the special configuration string by ":". The split needs to be avoided
when the ":" is inside some brackets.

Change-Id: I18ba7d4f7dcd7699cf6df3ebb0c8d3a33c78d5b0
Releases: 6.0
Resolves: #36727