[!!!][BUGFIX] Workspace state change notification email is always in English
The preferred language of the recipient is not taken into account which
bypasses the code used to notify user with a localized message in the
past. The end effect is that all notification messages are sent in
English.

Change-Id: I9ef3e12bf446e3786d7e582020adb75ba60dcd70
Depends: I7f87ad3b758f22fbc2259f7d73d8a15621fa9569
Fixes: #25434
Releases: 4.5, 4.6, 4.7, 6.0
Reviewed-on: http://review.typo3.org/12044
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers
Reviewed-by: Marco Bresch
Tested-by: Marco Bresch