[BUGFIX] Adjustments to FAL-enabled RTE image and link dialogues
When no folder is selected, use $BE_USER->getDefaultUploadFolder().
Use the id of the base script so that the same styles are applied to
the window.
When enabled, move the folder creation form to the bottom of the
window.
Honour readonly filemounts.
Slightly increase the default opening width of the image dialogue.

Change-Id: I710e60ffe00125874532137422a4ee461b8850e7
Releases: 6.0
Resolves: #36815