[BUGFIX] FLUIDTEMPLATE - Serialization of 'Closure' is not allowed
When using FLUIDTEMPLATE and COA_INT or USER_INT on the same
level and partials are used in the Fluid template, then there
will be Closures. COA_INT and USER_INT are going to serialize
the cObj, which fails when a Closure shall be processed.

Since the Closures have been collected in $cObj->contentObjects
the fix is to use a cloned version of the cObj which does not
contain these contentObjects anymore for COA_INT and USER_INT.

Change-Id: Ib3d69da33940c05708b29efa42cddf521af9014b
Fixes: #36820
Related: #32295
Releases: 6.0, 4.7, 4.6