[TASK] Remove unreachable cli code in typo3/init.php
This cli process code at the end of typo3/init.php could never be
reached since valid keys are already checked before and 'status' is
not registered as valid key. This code snipped is dead for a long time
now and did not work for ages. The information was semi useful and
undocumented. It is removed without substitution.

Change-Id: I97db67b18e414ca2b75ae9fcac2a1e892fad2a26
Resolves: #36936
Releases: 6.0