[BUGFIX] Regression of "better condition in versionOL of t3lib_tstemplate"
In the previous TYPO3 release, one if condition check of
t3lib_TStemplate::versionOL was changes / improved.
But this improvement causes a Fatal error on some TYPO3 installations.

In this method $GLOBALS['TSFE']->sys_page->versionOL is called.
The previous check is_object($GLOBALS['TSFE']) was replaces with
TYPO3_MODE === 'FE'. For most cases this will be correct.
But in eID mode, TYPO3_MODE is FE, too. And there is no full blown
Frontend available. This behavious will case the mentioned fatal error.

This patch replaces the check with a more specific way.

Change-Id: Ib3f9ebf355ee820e5144dd484d9a3a5e7708ebd4
Related: #31139, #25144
Fixes: #36981
Releases: 6.0, 4.7, 4.6, 4.5