[FEATURE] Allow camelCased class names in ext_autoload
Remove the hurdle that class names in ext_autoload (the key part of
the returned array) have to be lower cased.

The patch adds two tests, the first one expects that a case sensitive
class name in ext_autoload.php in actually lowercases before it is
written to the autoload cache file. It fails if the autoloader patch
is not applied.
The second tests that camel cased classes can be successfully autoloaded.

Change-Id: I2ccf916d43a288a520f6f89e9002bdc4df2da0b2
Resolves: #37110
Releases: 6.0