[TASK] Remove className option from services
When using services (sv), one had to add "classFile" to the info array when registering a service in order to find the entered className. As this is now completely taken care of by the autoloader, the option is obsolete and is now deprecated.

Note that once this is in, the documentation needs to be adjusted as well.

Change-Id: I74c8b859362fded6fe16a8344dbb18ce0db72c19
Branches: 6.0
Resolves: #37115