[TASK] Remove classFile option from services
When using services (sv), one had to add "classFile" to the info array when registering a service in order to find the entered className. As this is now completely taken care of by the autoloader, the option is obsolete and is now deprecated.

Note that once this is in, the documentation needs to be adjusted as well.

Change-Id: Ic8817c80f7646e071ba71fa498972b70115715c5
Branches: 6.0
Resolves: #37115