[CLEANUP] Make TCEforms sys_file header area localizable
When editing a sys_file database record - the metadata of a file - the header area is not localized, and the CSS styles for the preview image are inline-written. The change cleans this up by moving the labels in elf file and the CSS styles into the according t3skin CSS file.

Change-Id: I177a46f210cbe90cd142f5236c33a754999d7bca
Branches: 6.0
Resolves: #37116