[BUGFIX] Wrong DB fields usage in TCEmain
The fix for issue #35260 introduced a new bug in method
t3lib_TCEmain::getPreviousLocalizedRecordUid() by adding a
field in a query in the wrong place. In particular, this breaks
the localization of tt_content elements used with IRRE.

This patch fixes the problem by separating the list of fields
used for selecting and for sorting.

Change-Id: Idac10ded43e817f22df34e5156b424829d4601c1
Fixes: #37221
Releases: 6.0, 4.7, 4.6, 4.5
Reviewed-on: http://review.typo3.org/11292
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn