[BUGFIX] Wrong DB fields usage in TCEmain
The fix for issue #35260 introduced a new bug in method
t3lib_TCEmain::getPreviousLocalizedRecordUid() by adding a
field in a query in the wrong place. In particular, this breaks
the localization of tt_content elements used with IRRE.

This patch fixes the problem by separating the list of fields
used for selecting and for sorting.

Change-Id: I66101071a0320b7c2b1826c3422e7bf99671a122
Fixes: #37221
Releases: 6.0, 4.7, 4.6, 4.5
Reviewed-on: http://review.typo3.org/11250
Reviewed-by: Francois Suter
Tested-by: Francois Suter
Reviewed-by: Philipp Gampe
Reviewed-by: Tolleiv Nietsch
Tested-by: Tolleiv Nietsch