[BUGFIX] Wrong DB fields usage in TCEmain
The fix for issue #35260 introduced a new bug in method
t3lib_TCEmain::getPreviousLocalizedRecordUid() by adding a
field in a query in the wrong place. In particular, this breaks
the localization of tt_content elements used with IRRE.

This patch fixes the problem by separating the list of fields
used for selecting and for sorting.

Change-Id: If471cc8a00ef849e94a16f05f144b2985c530831
Fixes: #37221
Releases: 6.0, 4.7, 4.6, 4.5
Reviewed-on: http://review.typo3.org/11294
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn