[BUGFIX] Wrong DB fields usage in TCEmain
The fix for issue #35260 introduced a new bug in method
t3lib_TCEmain::getPreviousLocalizedRecordUid() by adding a
field in a query in the wrong place. In particular, this breaks
the localization of tt_content elements used with IRRE.

This patch fixes the problem by separating the list of fields
used for selecting and for sorting.

Change-Id: I1173a61a073cf3c0b9583a64259949ea688f30df
Fixes: #37221
Releases: 6.0, 4.7, 4.6, 4.5