[TASK] Bootstrap: Exit cli with valid keys if no key is given
If no cli key is given, cli_dispatch.phpsh exits with a generic
error message that a key must supplied to run the script. If a
key is given, but it is no registered cli script, the script
exits with a list of valid keys.

The patch removes the first test to also show the list of valid
keys if there is no key at all and adapts the error message a bit.

Change-Id: I4d356a1837218aaeb92ef85d9c5b8833b2737e08
Resolves: #37238
Releases: 6.0
Reviewed-on: http://review.typo3.org/11263
Reviewed-by: Francois Suter
Tested-by: Francois Suter
Reviewed-by: Oliver Klee
Tested-by: Susanne Moog
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn