[FEATURE] Bootstrap: Introduce shutdown and use in autoloader
The patch introduces a shutdown method to the bootstrap that
can be used to perform clean up and shutdown things at the
end of a script run. It is used in the most important entry
scripts like FE, mod.php, cli and backend.

First usage of shutdown() is the autoloader: If new classes are found
during script run, for example because they stick to the extbase naming
scheme and do not have a ext_autoload entry, the autoloader wrote a new
cache file for each class it found. Now, the autoloader remembers that
the cache file should be updated and delays of writing the cache file
until it is unregistered.
This way the full cache entry is only written once during shutdown.

Change-Id: Ibb84a7db65eb28e792c0b5ee9bedec00b29dbbb9
Resolves: #37268
Releases: 6.0