[BUGFIX] DB field names exceed max length with 'zzz_deleted'
MySQL allows a maximum of 64 characters for column and table names.
When prefixing them with 'zzz_deleted' this limit might be reached
by accident. (e.g. with
'zzz_deleted_tx_powermail_tellAFriendForm_receiverEmailContentField')

The patch avoids the overstepping 64 characters.

At the same time the prefixed and shortened field names are more
likely to have the same name, therefore the names are cut
from the beginning.

Change-Id: I7e377e9fe26894b1c6e6b80b1b787e88b5df6087
Fixes: #33580
Releases: 4.5, 4.6, 4.7, 6.0