[TASK] Add a comment for require(Bootstrap.php)
Bootstrap.php is included using "require" instead of "require_once" to
create a fatal error if Bootstrap.php is included from another place as
well (which would be a bug).

In order to keep people from changing this to "require_once" (as
recommended by the CGL), there should be comment explaining the
situation.

Change-Id: I39c88cae5af38bceafa5ca92561394a275152520
Resolves: #37447
Relates: #37383