[TASK] Add visibility to t3lib_tcemain member variables
The patch changes as much member variables as possible to "protected"
only the ones which are in use from the outside are left public.

The less we open up the more can be moved / refactored later. From what
I found even internal variables are used at the moment and as long as
there's no technical restriction this might continue.

Change-Id: I2bd7ee95d88a5a46aeb95094adee8b660aaedbca
Resolves: #37497
Releases: 6.0