[TASK] Increase readability of cli messages
Increasing the readability of cli messages when calling them on the
command line.
This improvement displays more human readable messages to a cli
user when calling the cli_dispatcher with wrong or no arguments
leaving out technical details like "array" and the array keys.

The keys are also sorted alphabetically

Change-Id: I62f41f333cf52e46836aeef105021eb571c09110
Fixes: #37713
Releases: 6.0