[TASK] Remove TYPO3_tables_script and handling
The definition of constant TYPO3_tables_script in localconf.php allows to
set a different path to the basic tca definitions than the default
t3lib/stddb/tables.php. This was never recommended and there is a
preferred way to overload the default file with a "extTablesDef" script.

Since the TYPO3_tables_script was probably never used by anyone,
is hidden, underdocumentated and not very useful, the functionality is
now removed without substitution.

Change-Id: Ie780f56c6ae31f581c286a959168b719b1106079
Releases: 6.0
Resolves: #38081