[FEATURE] Add module menu object to the TYPO3 backend
The backend at the moment uses a combination of the
GLOBAL variables TBE_MODULES and TBE_MODULES_EXT to
build the main menus of the backend (the module menu
on the left and the submodule menus of the backend
modules - look at info for an example).

To be able to easily change the representation of
these menus and to offer a clean API for rendering
menus the menu is refactored into an object structure
with n levels. This patch migrates the old GLOBALS
to a 3 level menu object. This means that you can use
the module menu object in your backend modules to
render your submenu.

This patch just adds the object structure and
transformation. It does not change any handling of
menus in the core itself. These changes can be done
piece by piece after this patch.

Change-Id: I1f7f9f5fcf83274840acef6cbeb267ea4ca460c3
Resolves: #38138
Releases: 6.0