[FEATURE] Add event handling to bootstrap mechanism
Adds the possibilities to register simple events that are called
once the the bootstrap reaches a given state (e.g. initialized).

NOTICE: This is an intermediate solution and not yet ready to be
fully reviewed. Feel free to add your comments here or to the
accordant issue on forge.typo3.org.

Change-Id: I1322b85f24a427aa03b05913d9f1edc5483875c0
Resolves: #38233
Releases: 6.0