[FEATURE] Introduce context based URL processing of FAL resources
If one would like to modify an URL of a file that is served by
an external FAL storage, this needs to be done for a given
rendering context. This change (re-)introduces a hook in
tslib_cObj that passes the accordant file and the context
TypoScript configuration to the callback definition.

Change-Id: I6a410399f34fa45ec03c959e23615407d451e4ae
Resolves: #38239
Releases: 6.0