[FEATURE] Add object representation for the Clipboard
The Clipboard Data is stored as a serialized multi-level array.
Where it is stored (wether User-Configuration or session) is determined
by by a User-TS setting.
All the logic how to deal with that data-structure resides in t3lib_clipboard
which only can be controlled via $_GET parameters.

This changeset adds two new Classes as object abstraction of the Clipboard
concept. It is fully compatible with the current clipboard and just reimplements
or extracts the logic to objects.

In a next step the old clipboard class will be adapted to use the new Objective
Interface.

Change-Id: I28c564763144a38fd0f4fb653fea81beb1854820
Releases: 6.0
Resolves: #32150