[BUGFIX] Missing class definitions in core autoloader
Since the classes in typo3/ have been moved to the TYPO3 Core
Autoloader as well and accordant require() calls removed, some
parts of the TYPO3 Backend will show a PHP Fatal Error like
"class not found".
The reason is, that there are PHP files with multiple class
definitions that have not been considered.

The changed autoload registry file was created with the SVN
version of EXT:extdeveval.

Change-Id: Id78e479ea51757eead7ec7298e8749c44d148ab0
Fixes: #38293
Releases: 6.0
Reviewed-on: http://review.typo3.org/12286
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel