[BUGFIX] CamelCased class names impossible with empty cache
Issue #37110 that aimed to allow case sensitive class name in
ext_autoload files has a bug that any class registered with CamelCased
class name is only found if the autoload entries are read from cache.

The patch now creates a lower cased entry as soon as the core_autoload
and ext_autoload files are read. There is an additional test that fails
if the changes in the autoload class are not applied. The naming of the
existing camelCase test was adapted to show that it is handling the
answer of an successful cache fetch only.

Additionally the current strtolower is changed to t3lib_div::strtolower
to make the whole thing multibyte safe.

Change-Id: I94f7ca5b3e60234f08433c747c00639d8dd219cd
Fixes: #38331
Related: #37110
Releases: 6.0