[TASK] Encapsulate bootstrap base code
This patch merges all bootstrap code that is independent from the local
instance and configuration into one class. A wrapper method in
Typo3_Bootsrap is introduced to call the only public method of this
class. This reduces the number of public bootstrap methods and
creates a clear definition what code must be run until instance
specific configuration comes into the game.

Some code fragments like the time initialization is now done more early
in bootstrap, the according methods are merged in the class.

The basic error_reporting setup is moved to the very end of the class,
which forced some called t3lib_div methods to be notice free.

All code that can be easily tested in this class now has unit tests.

Change-Id: I90d48f8f4ff4a59b5507becb6102c90b39d9eebd
Resolves: #38344
Releases: 6.0
Reviewed-on: http://review.typo3.org/12319
Reviewed-by: Steffen Ritter
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Susanne Moog
Tested-by: Susanne Moog