[!!!][TASK] Move temp_CACHED_* files to caching framework
After the basic bootstrap refactoring that moved the loading of
ext_tables.php and ext_localconf.php to class context, it is now easily
possible to put the former temp_CACHED_* files to the caching framework
as cached PHP code, next to the autoloader registry cache entry.

This patch introduces a new and clean API to t3lib_extMgm to handle the
loading of ext_localconf, ext_tables and the global TYPO3_LOADED_EXT
array and removes the current handling that was cluttered throughout
the core.

The setting of TYPO3_CONF_VARS['SYS']['extCache'] is now deprecated and
unused. The bootstrap now decides based on its context, whether or not
caching of those files is used or not. This way we can for example now
run the install tool in a fully uncached environment and can get rid of
the nasty side effects we had in this area in the past.

With a simple call to t3lib_extMgm::removeCacheFiles() all cache entries
in PHP code cache that are tagged with 'core' can be removed at once.
This currently affects the autoloader registry, the TYPO3_LOADED_EXT
entry, and the concatenated ext_localconf and ext_tables files.

The patch furthermore gets rid of the _CACHEFILE array key in
$GLOBALS['TYPO3_LOADED_EXT'] that worked as a kind of API to the calling
method to indicate if a concatenated ext_tables and ext_localconf file
is available.

Change-Id: Id00be5f24ec1698ed9126e61b229a718d6ca6741
Resolves: #38414
Releases: 6.0