[BUGFIX] Adding Multiple Images fails
Adding multiple images using the "import elements" link within
the record picker fails, since it calls a request to collapse
all already existing images of the element, without handing the
needed parameters.

As the needed parameters are not available at this point, just ommit
the request to close collapse the IRRE entries.

Change-Id: Ibc870cdcac32b4f5867ae12ca1020a20a5905310
Releases: 6.0
Resolves: #38383
Reviewed-on: http://review.typo3.org/12367
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn