[BUGFIX] Only set request for Web_Response
The current request is needed in Tx_Extbase_MVC_Web_Response to
distinguish between a cached or a non-cached request in
addAdditionalHeaderData() as the PageRenderer does not support USER_INTs
currently. This works fine is the ActionController is called with a
Web_Response, but causes an fatal error if a Cli_Response is in charge.

This fix and the $request member in Web_Response should be removed,
once the PageRenderer is fixed for USER_INTs. For now, this patch fixes
a fatal error in the fluid unit tests.

Change-Id: I56fabc3a51ce2c5192ab4488c817b938ed8797ca
Fixes: #36751
Releases: 6.0
Reviewed-on: http://review.typo3.org/12401
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel
Reviewed-by: Daniel Lorenz
Tested-by: Daniel Lorenz
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn