[BUGFIX] Bootstrap must check if submodules exist
A fresh clone of the git repository does not contain submodules.
That breaks bootstrap because some extensions require extbase to exist.

Check if the folders of all submodules are not empty during bootstrap
and provide a usefull error message if they are empty.

This patch only targets 6.0 due to the refactored bootstrap.

Change-Id: Ibbe1e0294cd5f9192895dfee3e7cee3dcddb6e48
Fixes: #38445
Releases: 6.0