[TASK] Autoloader performance improvement
If the autoload registry is read from cache, all class names are
already lowercased. The patch calls the lowercase logic only if
the registry files are not read from cache. The unit tests show
that nothing breaks with this patch.

Cachegrind shows ~22% of rendering time is spend in
loadCoreAndExtensionRegistry() on a fully cached page,
this is reduced to ~2% with the patch.

Change-Id: I07dc95b29bf05970061b5fb2f9a5d8a5ba194960
Resolves: #38498
Releases: 6.0
Reviewed-on: http://review.typo3.org/12494
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Susanne Moog
Tested-by: Susanne Moog