[TASK] Build more PHP modules in travis environment
In order to execute certain unit tests special requirements 
of the environment must be met. In particular several 
additional PHP modules need to be installed.

Instead of hiding what is added to the environment, 
use a helper script to not clutter the travis 
configuration file, but at the same time make it 
transparent there which PHP modules are built.

Change-Id: I7111cbd0f3d1d32d68df6eaf8c4b6176bfe82eff
Releases: 6.0