[TASK] Move buttons in user setting into admin functions
The buttons "clear configuration" and "reset
temporary data" are admin functions and should
therefore be moved to the admin tab.

Additionally they divert the attention, so that
it is very hard to notice the save button on the
top.

The patch moves the buttons to the admin tab and
additionally aligns the select box(es) with the
input and button fields.

Change-Id: I5c5552d3a20b68e458b6523bfb5e939000cd4f9a
Related: #632
Releases: 6.0
Resolves: #38523