[!!!][TASK] New local configuration handling
This patch removes the file typo3conf/localconf.php and introduces
a new file called typo3conf/LocalConfiguration.php instead. The file
returns the local overrides of the TYPO3_CONF_VARS default array,
which is merged with the default array during bootstrap.

An upgrade wizard transfers the settings from localconf to the new
file. As an effect, the LocalConfiguration file is now fully under
core control. There must be no additional custom code in it. If an
instance needs such code, it can be added to a second file called
typo3conf/AdditionalConfiguration.php, which is executed as is
during bootstrap after LocalConfiguration was read.

As an effect, this enables us to store the local configuration
as an array, exports the file with a clear schema, and enables
us to also store the list of enabled extensions as array.

Important notes:
- Frontend, backend, cli and install tool must still come up, even
  if the upgrade wizard was not run, yet.
- The variables '$typo3_db_*' that where in localconf.php are now
  merged with the TYPO3_CONF_VARS. The upgrade wizard should find
  and transfer them.
- The settings in install toll -> Basic Configuration are currently
  broken. This would have made the patch even bigger, and must be
  fixed with an additional patch.
- It is advised to run the update wizard as soon as possible after
  this patch is merged, it will be the first wizard in the line.
- Some follow ups for dbal and probably workspaces are still needed.
- phpunit fatals if the patch is applied, a patch is attached
  to the issue on forge as an immediate fix.
- Currently the 1-2-3 install wizards are broken. The dummy and
  other packages need some love to adapt to the new handling.

If some of the above problems are not solved until 6.0
stabilizes, this patch needs to be reverted again.

Change-Id: I3bf6a176117f501946123b921b6d2f1932627270
Resolves: #38562
Release: 6.0