[BUGFIX] In IE9, RTE does not work correctly in compat modes IE8/IE7
Problem: When using IE9, the RTE does not work correctly when using
compatibility modes IE8 or IE7. In particular, the style sheets are not
parsed, because they are not loaded in the same order as with the
specified compatibility version.
Check whether compatibility mode is activated and avoid
reliance on order.

Change-Id: I4891a15bf610633a75373c8ff36122b4ddd053f4
Releases: 4.5, 4.6, 4.7, 6.0
Resolves: #38574