[TASK] Always return a boolean in t3lib_div::validPathStr
In t3lib_div::validPathStr() a boolean (TRUE) is returned,
if this is a valid path string. If it is NOT a valid path string,
then nothing will be returned. In the doc comment,
there is a return type "boolean" mentioned.

Just return a boolean, if it is not a valid path string.

Change-Id: Ie1de33df3520cb0560d9f2c1a4bdd629bc897c6c
Fixes: #38604
Releases: 6.0, 4.7, 4.6, 4.5