[TASK] Always return a boolean in t3lib_div::validPathStr
In t3lib_div::validPathStr() a boolean (TRUE) is returned,
if this is a valid path string. If it is NOT a valid path string,
then nothing will be returned. In the doc comment,
there is a return type "boolean" mentioned.

Just return a boolean, if it is not a valid path string.

Change-Id: Iaf6f9fa0a8e3e414880b9b89d847c0be124a55d5
Fixes: #38604
Releases: 6.0, 4.7, 4.6, 4.5