[FEATURE] Include jQuery in the TYPO3 Backend
jQuery 1.8 (beta1 currently, final will be released in July 2012)
is now included in the typo3/contrib/ directory.
Additionally jQuery in a noConflict mode is available in the
TYPO3 Backend through "TYPO3.jQuery".
The page renderer also has methods to include jQuery versions
through CDNs and different versions.
A debug version and page renderer option will be added
at a later stage, once the final 1.8 version is released. Additionally
the support for jQuery Plugins and for TypoScript inclusion will be
added in a separate commit.

Change-Id: I659e50867a189fe4b7d5ac8313040dcecb58ee71
Resolves: #35428
Releases: 6.0