[FEATURE] Add jQuery API to the pagerenderer
jQuery 1.8 (beta1 currently, final will be released in July 2012)
is now included in the typo3/contrib/ directory.
Additionally jQuery in a noConflict mode is available in the
TYPO3 Backend through "TYPO3.jQuery".
The page renderer also has methods to include jQuery versions
through CDNs and different versions.

It also add namespace support, so that multiple versions of jQuery
can be included in different namespaces. It is also possible
to disable the namespace completely and let jQuery register itself
globally which can be useful for frontend rendering.

Once the final 1.8 version is released, the minified version of jQuery
should be included as well. Currently we included two files both with
the uncompressed version of jQuery 1.8 beta1

The support for jQuery Plugins and for TypoScript inclusion will be
added in separate commits.

Change-Id: I659e50867a189fe4b7d5ac8313040dcecb58ee71
Resolves: #38709