[TASK] Always return a boolean in t3lib_div::validPathStr
In t3lib_div::validPathStr() a boolean (TRUE) is returned,
if this is a valid path string. If it is NOT a valid path string,
then nothing will be returned. In the doc comment,
there is a return type "boolean" mentioned.

Just return a boolean, if it is not a valid path string.

Change-Id: Ie153a6df39f639895264dafbbd1adc76d96ae124
Fixes: #38604
Releases: 6.0, 4.7, 4.6, 4.5
Reviewed-on: http://review.typo3.org/12667
Reviewed-by: Andy Grunwald
Reviewed-by: Oliver Klee
Reviewed-by: Marcus Schwemer
Tested-by: Marcus Schwemer
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer