[FEATURE] Add categorization into the Core
A common use case in every advanced website is to be able to
categorize records. Besides that, there is the need to share
categories across records. So far, each extension has to
bring its own category implementation which is not an ideal
situation.

To fill the gap, we would like to introduce a new category record
type along with an API where extension developers could register
their own tables to be categorized. The relations will be stored
within "mm" tables as a less time consuming approach.

This patch provides:

* SQL definition for "sys_category" and "sys_category_mm" with their
TCA

* a registration mechanism where third party extension can have
their SQL fields + TCA generated on the fly. This is done by using
the Extension Manager method:

t3lib_extMgm::makeCategorizable($extensionKey, $tableName,
$fieldName = 'categories', $options = array())

Change-Id: I461252b6d5f6c6c4a4eb2c1942a66250cbb95aa9
Resolves: #38711
Releases: 6.0