[FEATURE] TYPO3 Webservice API
This provides an easy-to-use API which can be used by extension developers
to provide general webservices for their extensions. All in all it is highly
inspired by "DSL" PHP frameworks like Laravel, Slim and Silex which serve
with a way of defining URI/ACTION combinations kinda "inline"
(like $app->post('/users', function() { ..do stuff to add users.. }) ).

This is the base for a to-be-released rest extension which will provide a
RESTful Api for TYPO3. Because REST is all about hypermedia which should not
be the job of this core addition. Things like authentication, authorization,
database access, variable filtering, etc. are currently consciously held out
of this patch to observe the general utilization of this API. However, this
stuff can be done within user-defined classes within the lambda functions.

Currently provided features:
  - as thin as possible, just dispatch the requests and refer to
    EXT:<requestedExtension>/ext_api.php
  - differentiate between all the different HTTP methods
  - provide URI templates (like /user/:id)

Kudos to Thomas Maroschik who assembled the first few object fragments back
in 2011. For further details and discussion, refer to
http://wiki.typo3.org/TYPO3_Webservice_API.

Change-Id: I1e291fd2cae081fd29897d7c3408f95e56e40136