[FEATURE] TCEforms: Replace "field has changed" icon with CSS class
When editing a record in TCEforms, each field that is modified
gets an arrow symbol in front of the input/select field.
There are certain downsides to this:
- Having images and hard-coded srcs in each field clutters
the HTML and has ugly JS to change the field
- plus it's hard to style the icon
- The current idea does not work for FlexForm fields

The new solution adds a generic TCEforms jQuery object
that hooks into the existing fieldChanged function,
and does the replacement through CSS classes.
The CSS classes now, according to current design standards,
just adds a border around the modified image.

Change-Id: I94b6cc8550f7539961aa7ee97a8081e7d0ae5931
Resolves: #38752
Releases: 6.0