[FEATURE] Add signal to process public URL of file
By using the File Abstraction Layer (FAL), resources can be put
to remote storages like Amazon S3 or WebDAV. However, in those
cases a protected URI (containing user credentials) is used for
uploading which shall not be accessible for any user of course.

This change introduces the possibility to pre-process the public
URL generation and allows i.e. to substitute by an alternative
dispatcher URL.

Example for registering the slot:
$this->getSignalSlotDispatcher()->connect(
	't3lib_file_Storage',
	t3lib_file_Storage::SIGNAL_PreGeneratePublicUrl,
	'Tx_MyExtension_HandlerSlot',
	'preGeneratePublicUrl'
);

Change-Id: I9365831208dc90427aed29cff5672cc8ba5d6261
Resolves: #38821
Releases: 6.0
Reviewed-on: http://review.typo3.org/12732
Reviewed-by: Daniel Sattler
Tested-by: Daniel Sattler
Reviewed-by: Benjamin Mack
Tested-by: Benjamin Mack