[BUGFIX] Invalid shortcut target on translated pages
A feature to resolve shortcut links directly in menu rendering
does not consider the overlay behaviour of the "shortcut" value.

TSFE first uses the "shortcut" value of the original/default
page and then serves for possible translated alternatives
(tslib_fe::checkTranslatedShortcut()). In menu rendering it's
the other way round, the translated overlay (if any) is used
directly - the fallback to use the "shortcut" value of the
original/default page is not implemented.

This change introduces the fallback when rendering menus, which
will take the "shortcut" value of the default language if the
value in the overlay is empty.

Change-Id: I9bcbfec19ed3b98cc5794694f7a116f9ebdf6e90
Fixes: #36822
Releases: 6.0, 4.7, 4.6